Fix: Settings nav tab does nothing -> make it toggle the rail variant
