+++ samples/Flare.Gallery/Program.cs
@@ -60,7 +60,7 @@ builder.Services.AddFlareVersionCheck(opts =>
 builder.Services.AddSingleton<TimeProvider>(TimeProvider.System);
 builder.Services.AddLocalization();
 builder.Services.AddScoped<LanguageService>();
-builder.Services.AddScoped<DrawerModeService>();
+builder.Services.AddScoped<RailLabelService>();
 builder.Services.AddSingleton<GallerySearchService>();
 builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
 
@@ -70,9 +70,9 @@ var host = builder.Build();
 var languageService = host.Services.GetRequiredService<LanguageService>();
 await languageService.InitializeCultureAsync();
 
-// Restore the saved navigation drawer mode before first paint so the layout opens in the user's
-// preferred mode (no flash from the default).
-var drawerModeService = host.Services.GetRequiredService<DrawerModeService>();
-await drawerModeService.InitializeAsync();
+// Restore the saved rail-label preference before first paint so the rail renders in the user's
+// preferred variant (no flash from the default).
+var railLabelService = host.Services.GetRequiredService<RailLabelService>();
+await railLabelService.InitializeAsync();
 
 await host.RunAsync();
